Rename index to SQLClient and SQLClient to SQL
